app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
python-multipart = "^0.0.5"
email-validator = "^2.1.0.post1"
requests = "^2.31.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tenacity = "^8.2.3"
emails = "^0.5.15"
raven = "^6.10.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
pytest = "^7.4.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlmodel = "^0.0.11"
uvicorn = "^0.24.0.post1"
pypandoc = "^1.12"
slowapi = "^0.1.9"
dulwich = "^0.21.7"

[tool.poetry.dev-dependencies]
mypy = "^1.8.0"
black = "^24.1.0"
isort = "^5.13.2"
autoflake = "^2.2.1"
flake8 = "^7.0.0"
pytest = "^7.4.4"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^4.1.0"

[tool.poetry.group.dev.dependencies]
locust = "^2.19.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

